Biryani classes Chodechef problem
Problem
According to a recent survey, Biryani is the most ordered food. Chef wants to learn how to make world-class Biryani from a MasterChef. Chef will be required to attend the MasterChef's classes for XX weeks, and the cost of classes per week is YY coins. What is the total amount of money that Chef will have to pay?

Input Format
The first line of input will contain an integer TT — the number of test cases. The description of TT test cases follows.
The first and only line of each test case contains two space-separated integers XX and YY, as described in the problem statement.
Output Format
For each test case, output on a new line the total amount of money that Chef will have to pay.

Constraints
1 \leq T \leq 10^41≤T≤10 
4
 
1 \leq X, Y \leq 1001≤X,Y≤100
Sample 1:
Input
Output
4
1 10
1 15
2 10
2 15
10
15
20
30
Explanation:
Test case 11: Chef will be required to attend the MasterChef's classes for 11 week and the cost of classes per week is 1010 coins. Hence, Chef will have to pay 1010 coins in total.

Test case 22: Chef will be required to attend the MasterChef's classes for 11 week and the cost of classes per week is 1515 coins. Hence, Chef will have to pay 1515 coins in total.

Test case 33: Chef will be required to attend the MasterChef's classes for 22 weeks and the cost of classes per week is 1010 coins. Hence, Chef will have to pay 2020 coins in total.

Test case 44: Chef will be required to attend the MasterChef's classes for 22 weeks and the cost of classes per week is 1515 coins. Hence, Chef will have to pay 3030 coins in total.