Audible Range
Problem
Chef's dog binary hears frequencies starting from 6767 Hertz to 4500045000 Hertz (both inclusive).

If Chef's commands have a frequency of XX Hertz, find whether binary can hear them or not.

Input Format
The first line of input will contain a single integer TT, denoting the number of test cases.
Each test case consists of a single integer XX - the frequency of Chef's commands in Hertz.
Output Format
For each test case, output on a new line YES, if binary can hear Chef's commands. Otherwise, print NO.

The output is case-insensitive. Thus, the strings YES, yes, yeS, and Yes are all considered the same.

Constraints
1 \leq T \leq 10^41≤T≤10 
4
 
1 \leq X \leq 10^61≤X≤10 
6
 
Sample 1:
Input
Output
5
42
67
402
45000
45005
NO
YES
YES
YES
NO
Explanation:
Test case 11: Chef's command has a frequency of 4242 Hertz which is less than 6767. Thus, it would not be audible to binary.

Test case 22: Chef's command has a frequency of 6767 Hertz which lies in the range [67, 45000][67,45000]. Thus, it would be audible to binary.

Test case 33: Chef's command has a frequency of 402402 Hertz which lies in the range [67, 45000][67,45000]. Thus, it would be audible to binary.

Test case 44: Chef's command has a frequency of 4500045000 Hertz which lies in the range [67, 45000][67,45000]. Thus, it would be audible to binary.

Test case 55: Chef's command has a frequency of 4500545005 Hertz which is greater than 4500045000. Thus, it would not be audible to binary.